POMP-78: Refactoring for greater modularity
Created a new Module in shared for loading services, to decouple modules as far as possible. Moved mapping api to new folder and changed it module. Concrete implementations will be placed in the top level mapping module as well. Next step is to remove Model to ProjectDTO mapping to a own mapping module and restructure it in more testable classes.