POMP-81: Adding more unit tests for DirectedGraph
Noticed incosistency with methods which delete all objects within a collection from the graph. Should All elements which can be deleted be deleted and the exceptions be collected, or should all changes be rolled back, when one element does throw an exception, for Atomicity and Consistency? This has to be determined

Took 1 hour 35 minutes